Big refactor, mostly to support getting all pages instead of the first one
Looked for an elegant solution for a long time. The first option was to use Response in leagues.go, series.go and matches.go to continuously fetch all pages but that meant that this logic would have to be copied into each of those files, which ain't no good.

This logic seemed to belong in execution.go but I couldn't get it to work using structs. This seems to require reflect, which I really wanted to avoid. The best option I've found now is to unmarshall and append into []map[string]interface{} and then marshall/unmarshall that map into the pointer to the struct.